refactor: separate mocks by type (service/repository) for section and product batch
